+++ src/state/createStore.js
@@ -1,6 +1,8 @@
 import { applyMiddleware, combineReducers, compose, createStore } from 'redux';
 import createSagaMiddleware from 'redux-saga';
-import gearOptimizerSaga from './optimizer/sagas';
+import gearOptimizerSaga from './sagas/sagas';
+import formStateSaga from './sagas/formStateSaga';
+import buildPageSaga from './sagas/buildPageSaga';
 import { bossSlice } from './slices/boss';
 import { buffsSlice } from './slices/buffs';
 import { buildPageSlice } from './slices/buildPage';
@@ -43,5 +45,7 @@ const composeEnhancers =
 export default () => {
   const store = createStore(reducers, composeEnhancers(applyMiddleware(saga)));
   saga.run(gearOptimizerSaga);
+  saga.run(formStateSaga);
+  saga.run(buildPageSaga);
   return store;
 };
