+++ src/state/sagas/sagas.js
@@ -1,8 +1,5 @@
 /* eslint-disable no-console */
-import { decompress } from '@discretize/object-compression';
-import JsonUrl from 'json-url';
-import { channel } from 'redux-saga';
-import { all, call, cancelled, put, race, select, take, takeLeading } from 'redux-saga/effects';
+import { all, call, cancelled, put, race, select, take } from 'redux-saga/effects';
 import {
   mapEntries,
   mapValues,
@@ -10,11 +7,10 @@ import {
   parseInfusionCount,
   parsePriority,
 } from '../../utils/usefulFunctions';
+import { calculate } from '../optimizer/multiCoreHandler';
+import { ERROR, SUCCESS, WAITING } from '../optimizer/status';
 import { getBuffsModifiers } from '../slices/buffs';
-import { changeBuildPage } from '../slices/buildPage';
-// import { changeBuildPage } from '../slices/buildPage';
 import {
-  changeAll,
   changeControl,
   changeError,
   changeFilteredList,
@@ -29,9 +25,6 @@ import { getInfusionsModifiers } from '../slices/infusions';
 import { getCustomAffixData } from '../slices/priorities';
 import { getSkillsModifiers } from '../slices/skills';
 import { getCurrentSpecialization, getTraitsModifiers } from '../slices/traits';
-import { calculate } from './multiCoreHandler';
-import SagaTypes from './sagaTypes';
-import { ERROR, SUCCESS, WAITING } from './status';
 
 const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
 
@@ -291,183 +284,6 @@ function* watchStart() {
   }
 }
 
-const lib = JsonUrl('lzma');
-
-const modifyState = (optimizerState) => {
-  // const list = optimizerState?.control?.list.slice(0, 6) || [];
-  // let modifiedList = list;
-
-  // // extract settings object from characters (should be identical)
-  // let listSettings = null;
-  // if (list[0]?.settings) {
-  //   listSettings = list[0].settings;
-  //   modifiedList = list.map((character) => {
-  //     // eslint-disable-next-line no-unused-vars
-  //     const { settings, ...rest } = character;
-  //     return { ...rest };
-  //   });
-  // }
-
-  let modifiedSelectedCharacter = null;
-  if (optimizerState.control.selectedCharacter) {
-    modifiedSelectedCharacter = JSON.parse(
-      JSON.stringify(optimizerState.control.selectedCharacter),
-    );
-    modifiedSelectedCharacter.settings.appliedModifiers = [];
-  }
-
-  const exportData = {
-    // listSettings,
-    optimizerState: {
-      ...optimizerState,
-      control: {
-        ...optimizerState.control,
-        list: [],
-        filteredList: [],
-        saved: [],
-        selectedCharacter: modifiedSelectedCharacter,
-      },
-    },
-  };
-
-  return exportData;
-};
-
-const unModifyState = (importData) => {
-  // eslint-disable-next-line no-unused-vars
-  const { optimizerState, listSettings } = importData;
-
-  // // replace settings objects in characters
-  // if (Array.isArray(optimizerState?.control?.list) && listSettings) {
-  //   optimizerState.control.list = optimizerState.control.list.map((character) => ({
-  //     ...character,
-  //     settings: listSettings,
-  //   }));
-  // }
-
-  return optimizerState;
-};
-
-function* exportState({ onSuccess }) {
-  const reduxState = yield select();
-
-  const exportData = modifyState(reduxState.optimizer);
-  console.log(exportData);
-
-  console.time('Created template in:');
-  const compressed = yield lib.compress(exportData);
-  console.timeEnd('Created template in:');
-
-  onSuccess(compressed);
-}
-
-function* watchExportState() {
-  yield takeLeading(SagaTypes.ExportFormState, exportState);
-}
-
-function* importState({ buildUrl: input, onSuccess, onError }) {
-  try {
-    if (!input) return;
-
-    console.time('Decompressed template in:');
-    const importData = yield lib.decompress(input);
-    console.timeEnd('Decompressed template in:');
-
-    console.log(importData);
-    const optimizerState = unModifyState(importData);
-
-    console.time('Applied state in:');
-    yield put(changeAll(optimizerState));
-    console.timeEnd('Applied state in:');
-
-    // execute success callback
-    onSuccess();
-  } catch (e) {
-    console.log('Problem restoring template!');
-    console.log(e);
-    onError();
-  }
-}
-
-function* watchImportState() {
-  yield takeLeading(SagaTypes.ImportFormState, importState);
-}
-
-/*
-function* exportStateCharacter({ onSuccess }) {
-  const reduxState = yield select();
-
-  const { selectedLines: lines, selectedTraits: selected } = reduxState.optimizer.form.traits;
-  const { weapons, skills, character } = reduxState.optimizer.buildPage;
-
-  const exportData = {
-    character,
-    weapons,
-    skills,
-    traits: {
-      lines,
-      selected,
-    },
-  };
-  console.log(exportData);
-
-  console.time('Created template in:');
-  const compressed = yield lib.compress(exportData);
-  console.timeEnd('Created template in:');
-
-  onSuccess(compressed);
-}
-
-function* watchExportStateCharacter() {
-  yield takeLeading('EXPORT_STATE_CHARACTER', exportStateCharacter);
-}
-*/
-
-// channels solve the problem "how to get a value out of a callback"
-// https://stackoverflow.com/questions/43031832/how-to-yield-put-in-redux-saga-within-a-callback
-const decompressChannel = channel();
-function* importStateCharacter({ buildUrl: input, version }) {
-  if (!input) {
-    console.error('SAGA: No url parameter supplied');
-    return;
-  }
-  if (typeof version === 'undefined') {
-    console.error('SAGA: No version parameter supplied');
-    return;
-  }
-
-  try {
-    // load build state from url
-    const { BuildPageSchema: schema } = yield import(
-      `../../components/url-state/schema/BuildPageSchema_v${version}`
-    );
-
-    decompress({
-      string: input,
-      schema,
-      onSuccess: (result) => decompressChannel.put({ type: 'STATE_DECOMPRESS_FINISHED', result }),
-    });
-
-    const { result } = yield take(decompressChannel);
-
-    yield put(changeBuildPage(result));
-  } catch (e) {
-    console.log('Problem restoring template!');
-    console.log(e);
-  }
-}
-
-function* watchImportStateCharacter() {
-  yield takeLeading(SagaTypes.ImportBuildPageState, importStateCharacter);
-}
-
 export default function* rootSaga() {
-  yield all([
-    // other sagas go here
-    watchStart(),
-    watchExportState(),
-    watchImportState(),
-    // watchExportStateCharacter(),
-    watchImportStateCharacter(),
-  ]);
+  yield all([watchStart()]);
 }
